requirements.txt
snowflake-connector-python>=3.0.0
pandas>=2.0.0
plotly>=5.0.0
streamlit>=1.30.0
python-dotenv>=1.0.0
